Create mixins.py file and remove redundant code
